ai-tutor/pyproject.toml
[project]
name = "chainlit-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.302",
    "openai-agents>=0.0.7",
]
